backend/requirements.txt
# Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Caching==2.0.2

# Database
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7  # For PostgreSQL
mysqlclient==2.2.0      # For MySQL

# ML and Image Processing
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# File Processing
python-magic==0.4.27
werkzeug==2.3.7

# API Documentation
flask-restx==1.1.0
swagger-ui-bundle==0.0.9

# Monitoring and Logging
prometheus-client==0.17.1
structlog==23.1.0
python-json-logger==2.0.7

# Email and Notifications
flask-mail==0.9.1
twilio==8.6.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
redis==4.6.0
celery==5.3.1

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
responses==0.23.3

# Development
black==23.7.0
flake8==6.1.0
autopep8==2.0.2
pre-commit==3.3.3
